refactor: move Reference and FileReference to core-types
