+++ packages/core-types/src/ICoreIdHelper.ts
@@ -1,0 +1,9 @@
+import { CoreId } from "./CoreId";
+
+export interface ICoreIdHelper {
+    readonly prefix: string;
+
+    generate(): Promise<CoreId>;
+    generateUnsafe(): Promise<CoreId>;
+    validate(id: string | CoreId): boolean;
+}
