+++ packages/core-types/src/SharedPasswordProtection.ts
@@ -1,6 +1,6 @@
 import { ISerializable, Serializable, serialize, validate } from "@js-soft/ts-serval";
 import { CoreBuffer, ICoreBuffer } from "@nmshd/crypto";
-import { TransportCoreErrors } from "../TransportCoreErrors";
+import { CoreError } from "../CoreError";
 
 export interface ISharedPasswordProtection extends ISerializable {
     passwordType: "pw" | `pin${number}`;
@@ -25,7 +25,7 @@ export class SharedPasswordProtection extends Serializable implements ISharedPas
 
         const splittedPasswordParts = value.split("&");
         if (splittedPasswordParts.length !== 2) {
-            throw TransportCoreErrors.general.invalidTruncatedReference("The password part of a TruncatedReference must consist of exactly 2 components.");
+            throw new CoreError("error.core-types.invalidTruncatedReference", "The password part of a TruncatedReference must consist of exactly 2 components.");
         }
 
         const passwordType = splittedPasswordParts[0] as "pw" | `pin${number}`;
@@ -33,7 +33,7 @@ export class SharedPasswordProtection extends Serializable implements ISharedPas
             const salt = CoreBuffer.fromBase64(splittedPasswordParts[1]);
             return SharedPasswordProtection.from({ passwordType, salt });
         } catch (_) {
-            throw TransportCoreErrors.general.invalidTruncatedReference("The salt needs to be a Base64 value.");
+            throw new CoreError("error.core-types.invalidTruncatedReference", "The salt needs to be a Base64 value.");
         }
     }
