+++ packages/core-types/src/index.ts
@@ -2,3 +2,6 @@ export * from "./CoreAddress";
 export * from "./CoreDate";
 export * from "./CoreError";
 export * from "./CoreId";
+export * from "./ICoreIdHelper";
+export * from "./references";
+export * from "./SharedPasswordProtection";
