+++ packages/core-types/src/references/FileReference.ts
@@ -1,12 +1,12 @@
 import { type } from "@js-soft/ts-serval";
-import { BackboneIds, IReference, Reference } from "../../../core";
+import { IReference, Reference } from "./Reference";
 
 export interface IFileReference extends IReference {}
 
 @type("FileReference")
 export class FileReference extends Reference implements IFileReference {
     protected static override preFrom(value: any): any {
-        super.validateId(value, BackboneIds.file);
+        // super.validateId(value, BackboneIds.file);
 
         return value;
     }
