+++ packages/core-types/src/references/index.ts
@@ -1,0 +1,2 @@
+export * from "./FileReference";
+export * from "./Reference";
