+++ packages/runtime/src/useCases/transport/relationshipTemplates/CreateTokenForOwnRelationshipTemplate.ts
@@ -1,11 +1,10 @@
 import { Result } from "@js-soft/ts-utils";
-import { CoreAddress, CoreDate, CoreId } from "@nmshd/core-types";
+import { CoreAddress, CoreDate, CoreId, SharedPasswordProtection } from "@nmshd/core-types";
 import {
     AccountController,
     PasswordProtectionCreationParameters,
     RelationshipTemplate,
     RelationshipTemplateController,
-    SharedPasswordProtection,
     TokenContentRelationshipTemplate,
     TokenController
 } from "@nmshd/transport";
