+++ packages/transport/src/core/types/index.ts
@@ -1,5 +1,4 @@
 export * from "./CoreHash";
 export * from "./PasswordProtection";
 export * from "./PasswordProtectionCreationParameters";
-export * from "./SharedPasswordProtection";
 export * from "./TransportVersion";
