+++ packages/transport/src/modules/tokens/transmission/TokenContentRelationshipTemplate.ts
@@ -1,7 +1,6 @@
 import { ISerializable, Serializable, serialize, type, validate } from "@js-soft/ts-serval";
-import { CoreAddress, CoreId, ICoreAddress, ICoreId } from "@nmshd/core-types";
+import { CoreAddress, CoreId, ICoreAddress, ICoreId, ISharedPasswordProtection, SharedPasswordProtection } from "@nmshd/core-types";
 import { CryptoSecretKey, ICryptoSecretKey } from "@nmshd/crypto";
-import { ISharedPasswordProtection, SharedPasswordProtection } from "../../../core/types/SharedPasswordProtection";
 
 export interface ITokenContentRelationshipTemplate extends ISerializable {
     templateId: ICoreId;
